requirements.txt
google-api-python-client==1.7.4
oauth2client==4.1.3
google-cloud-secret-manager==1.0.0